requirements.txt
alabaster==0.7.9
Babel==2.3.4
click==6.7
docutils==0.13.1
Flask==0.12
imagesize==0.7.1
itsdangerous==0.24
Jinja2==2.9.2
MarkupSafe==0.23
py==1.4.32
py-postgresql==1.2.1
Pygments==2.1.3
pytest==3.0.5
pytz==2016.10
requests==2.12.4
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.5.1
Werkzeug==0.11.15
